requirements.txt
aioredis==1.3.1
amqp==1.4.9
anyjson==0.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
bcrypt==3.1.7
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
crypto==1.4.1
cryptography==2.8
Django==3.2.5
django-celery==3.3.1
django-cors-headers==3.2.1
django-redis==4.11.0
gitdb==4.0.7
GitPython==3.1.17
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
IPy==1.0
kombu==3.0.37
msgpack==0.6.2
Naked==0.1.31
paramiko==2.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
PyHamcrest==2.0.2
PyNaCl==1.3.0
pyOpenSSL==19.1.0
python-nmap==0.6.1
pytz==2019.3
PyYAML==5.3.1
redis==2.10.6
requests==2.23.0
service-identity==18.1.0
shellescape==3.8.1
six==1.14.0
smmap==4.0.0
TestLink-API-Python-client==0.8.1
txaio==20.4.1
urllib3==1.25.1
loguru==0.5.3
jsonpath==0.82
channels==3.0.3
psycopg2-binary==2.9.1
pyVmomi==7.0.2
fabric==2.6.0
